Added a gcloud based dataloader class.

+++ src/jaxpt/models/rope_gpt.py
@@ -55,7 +55,7 @@ class RoPE_GPT(nnx.Module):
         # pre-calculate the RoPE thetas
         query_size = config.n_embed // config.n_head
         base_freq = config.rope_base_freq**(2/query_size)
-        omega = jnp.ones((1, query_size)) * base_freq
+        omega = jnp.ones((1, query_size), dtype=config.dtype) * base_freq
         pow = jnp.arange(0, query_size)
         omega = jnp.repeat(omega**pow, config.block_size, axis=0)
         pos = jnp.arange(0, config.block_size)
